+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHK2NRSMNNCD7179KCNFHDVB"
+        "01KHKCN1YXSVN4RZ47EFVC8CBP"
     )
     r"""The unique identifier of the data source"""