+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHK2NRSGW00E6NSKBRZ3MAJ6"
+        "01KHKCN1YTHJ63GZZPEH7ST6F8"
     )
     r"""The unique identifier of the data source"""