+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHK2NRSJDZS5AK8KAZFMJ2N4"
+        "01KHKCN1YWEYWE4K1FH7MCGJD1"
     )
     r"""The unique identifier of the data source"""