+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHK2NRSR6EC3AS7MM0S3CREP"
+        "01KHKCN1Z0YKZ251XT1G1J9HQH"
     )
     r"""The unique identifier of the data source"""