+++ packages/orq-rc/RELEASES.md
@@ -5689,3 +5689,13 @@ Based on:
 - [python v4.4.0-rc.0] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.0] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.0 - packages/orq-rc
+
+## 2026-02-16 14:14:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.718.0 (2.824.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.4.0-rc.1] packages/orq-rc
+### Releases
+- [PyPI v4.4.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.1 - packages/orq-rc